requirements.txt
azure-functions
debugpy
requests
pyodbc
aiohttp